# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Blocksync",
        "version": "1.0.0",
        "description": ""
    },
    "host": "",
    "basePath": "",
    "schemes": ["http"],
    "paths": {
...
        "/storage/store": {
            "post": {
                "description": "",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "name": {
                                    "example": "any"
                                },
                                "contentType": {
                                    "example": "any"
                                },
                                "data": {
                                    "example": "any"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/storage/retrieve/{cid}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "cid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/entity/byId/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/entity/byOwnerAddress/{address}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/entity/collectionById/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/entity/collectionsByOwnerAddress/{address}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "address",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/entity/owner/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/tokenclass/contractaddress/{contractAddress}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "contractAddress",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/tokenclass/name/{name}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/tokenclass/class/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/token/name/{name}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "name",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/token/id/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/token/entity/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/token/collection/{id}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bonds/listBonds": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bonds/listBondsFiltered": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bonds/getByBondDid/{bondDid}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "bondDid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bonds/getPriceHistoryByBondDid/{bondDid}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "bondDid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bonds/getByBondCreatorDid/{creatorDid}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "creatorDid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bonds/getAccountBonds/{account}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "account",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/bond/get/outcomepayments/{bonddid}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "bonddid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "size",
                        "in": "query",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
        "/api/project/getByProjectDid/{projectDid}": {
            "get": {
                "description": "",
                "parameters": [
                    {
                        "name": "projectDid",
                        "in": "path",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    }
}